requirements.txt
streamlit==1.32.0
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
xlsxwriter==3.1.9
python-dotenv==1.0.1
chardet==5.2.0
openpyxl==3.1.2
readability-lxml==0.8.1
textstat==0.7.3
pillow==10.2.0
matplotlib==3.8.2
seaborn==0.13.2